Handle buffer local process environments
Use the buffer local environment when launching a shell.  This is done
by setting the process env in the traditional way using `setenv` then
copying that to the global *default* `process-environment` variable.
We have to store the original state of the global process-environment
and the value of the buffer local process-environment.

Once we have launched or hit an error, revert the
`process-environment` for both global and local values to the values
saved before we modified anything.

This means that any buffer local process-environment will be inherited
into the shell, so depending on the mode/hooks's enabled the user may
experience differing behaviour when launching a shell from different
buffers.